+++ src/astrameter/main.py
@@ -123,6 +123,10 @@ async def run_device(
 
     device: CT002 | Shelly
 
+    global_dedupe_time_window = cfg.getfloat(
+        "GENERAL", "DEDUPE_TIME_WINDOW", fallback=0.0
+    )
+
     if device_type in ["ct002", "ct003"]:
         ct_section = get_ct_section(device_type, cfg)
         ct_type = "HME-4" if device_type == "ct002" else "HME-3"
@@ -130,7 +134,7 @@ async def run_device(
         ct_udp_port = cfg.getint(ct_section, "UDP_PORT", fallback=UDP_PORT)
         wifi_rssi = cfg.getint(ct_section, "WIFI_RSSI", fallback=-50)
         dedupe_time_window = cfg.getfloat(
-            ct_section, "DEDUPE_TIME_WINDOW", fallback=0.0
+            ct_section, "DEDUPE_TIME_WINDOW", fallback=global_dedupe_time_window
         )
         consumer_ttl = cfg.getint(ct_section, "CONSUMER_TTL", fallback=120)
         debug_status = cfg.getboolean(ct_section, "DEBUG_STATUS", fallback=False)
@@ -260,22 +264,42 @@ async def run_device(
     elif device_type == "shellypro3em_old":
         logger.debug("Shelly Pro 3EM Settings:")
         logger.debug(f"Device ID: {device_id}")
-        device = Shelly(powermeters=powermeters, device_id=device_id, udp_port=1010)
+        device = Shelly(
+            powermeters=powermeters,
+            device_id=device_id,
+            udp_port=1010,
+            dedupe_time_window=global_dedupe_time_window,
+        )
 
     elif device_type == "shellypro3em_new":
         logger.debug("Shelly Pro 3EM Settings:")
         logger.debug(f"Device ID: {device_id}")
-        device = Shelly(powermeters=powermeters, device_id=device_id, udp_port=2220)
+        device = Shelly(
+            powermeters=powermeters,
+            device_id=device_id,
+            udp_port=2220,
+            dedupe_time_window=global_dedupe_time_window,
+        )
 
     elif device_type == "shellyemg3":
         logger.debug("Shelly EM Gen3 Settings:")
         logger.debug(f"Device ID: {device_id}")
-        device = Shelly(powermeters=powermeters, device_id=device_id, udp_port=2222)
+        device = Shelly(
+            powermeters=powermeters,
+            device_id=device_id,
+            udp_port=2222,
+            dedupe_time_window=global_dedupe_time_window,
+        )
 
     elif device_type == "shellyproem50":
         logger.debug("Shelly Pro EM 50 Settings:")
         logger.debug(f"Device ID: {device_id}")
-        device = Shelly(powermeters=powermeters, device_id=device_id, udp_port=2223)
+        device = Shelly(
+            powermeters=powermeters,
+            device_id=device_id,
+            udp_port=2223,
+            dedupe_time_window=global_dedupe_time_window,
+        )
 
     else:
         raise ValueError(f"Unsupported device type: {device_type}")